Add game engine example to Run view
