+++ application.fam
@@ -6,6 +6,13 @@ App(
     stack_size=4 * 1024,
     fap_icon="app.png",
     fap_category="GPIO",
-    fap_icon_assets="assets",
     fap_description="Open World Multiplayer game, best played with the VGM."
+    fap_icon_assets="assets",
+    fap_file_assets="assets", # Do not touch this and the next line, it is needed to generate sprites
+    fap_extbuild=(
+        ExtFile(
+            path="${FAP_SRC_DIR}/assets",
+            command="${PYTHON3} ${FAP_SRC_DIR}/engine/scripts/sprite_builder.py ${FAP_SRC_DIR.abspath}/sprites ${TARGET.abspath}/sprites",
+        ),
+    ),
 )
