+++ callback/callback.h
@@ -2,5 +2,14 @@
 #include <flip_world.h>
 #include <flip_storage/storage.h>
 
+//
+#include <furi.h>
+#include "engine/engine.h"
+#include "engine/game_engine.h"
+#include "engine/game_manager_i.h"
+#include "engine/level_i.h"
+#include "engine/entity_i.h"
+
 void free_all_views(void *context, bool free_variable_item_list);
+bool flip_world_custom_event_callback(void *context, uint32_t index);
 void callback_submenu_choices(void *context, uint32_t index);
